restructring code for cluster makeup
